+++ apps/site/app/components/BackHomeLink/styles/index.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import styled from '@wuh.site/components/styled'
+import styled from 'styled-components'
 
 export const Wrapper = styled.span`
   display: inline-flex;