+++ apps/site/app/components/TypewriterMotto/styles.ts
@@ -1,4 +1,4 @@
-import styled from '@wuh.site/components/styled'
+import styled from 'styled-components'
 
 export const Container = styled.div`
   position: relative;