+++ apps/site/app/guestbook/GuestbookPageView/styles/index.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import styled from '@wuh.site/components/styled'
+import styled from 'styled-components'
 
 export const PageWrapper = styled.div`
   max-width: 720px;